ride_info() works mostly
still some errors; only 680 rides have been logged. haven't double-checked that to park inventory.  rides that don't have a section titled "track" or "tracks" error out - this can be future rides where nothing is known and someone has just taken a picture of a track section in a parking lot that has been delivered to a park recently.  so, not much is lost.